IoT-Smart-Devices-connection-with-Nervousnet
This is a yet new manner to interact with the Nervousnet Project (http://www.nervousnet.ethz.ch/). With this code, you are able to use Arduino+Wifi+Sensor to save one sensor Data to the backbone of Nervousnet (in this case, it's a Accelerometer sensor).

This code is composed by a file to be flashed into the microcontroller (in this case, Arduino), a Java Server providing the web interface between the microcontroller and the Nervousnet database, and the Nervousnet mysql DB.